+++ packages/stream_chat_flutter/test/src/mocks.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:stream_chat_flutter/src/video/vlc/vlc_manager_desktop.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 class MockClient extends Mock implements StreamChatClient {
@@ -28,6 +27,8 @@ class MockChannel extends Mock implements Channel {
     when(deleteDraft).thenAnswer((_) async => EmptyResponse());
     when(() => deleteDraft(parentId: any(named: 'parentId'))).thenAnswer((_) async => EmptyResponse());
     when(() => currentUserLastMessageAtStream).thenAnswer((_) => Stream.value(null));
+    when(() => isGroup).thenReturn(false);
+    when(() => isOneToOne).thenReturn(false);
   }
 
   @override
@@ -114,8 +115,6 @@ class MockUser extends Mock implements User {}
 class MockOwnUser extends Mock implements OwnUser {}
 
 class MockAttachment extends Mock implements Attachment {}
-
-class MockVlcManagerDesktop extends Mock implements VlcManagerDesktop {}
 
 class MockStreamMemberListController extends Mock implements StreamMemberListController {
   @override